requirements.txt
kafka-python
opencv-python
numpy
tensorflow
Pillow
matplotlib
